+++ docs/demos/multi_page_basics/pages/redirect.py
@@ -4,10 +4,9 @@ from dash import html
 
 dash.register_page(
     __name__,
-    name="Analytic Apps",
     description="Welcome to my app",
     redirect_from=["/old-home-page", "/v2"],
     extra_template_stuff="yup",
 )
 
-layout = html.Div("Home Page")
+layout = html.Div(["Home Page", html.Img(src="/assets/birds.jpeg", height="50px")])