+++ docs/demos/multi_page_no_pages_folder/app.py
@@ -4,7 +4,7 @@ import dash_labs as dl
 
 app = Dash(__name__, plugins=[dl.plugins.pages])
 
-dash.register_page("another_home", layout="We're home!", path="/")
+dash.register_page("home", layout="We're home!", path="/")
 dash.register_page(
     "very_important", layout="Don't miss it!", path="/important", order=0
 )